ejercicio semiperimetro

+++ guias/triSemi.c
@@ -3,7 +3,7 @@
 
 int main() {
     int a, b, c;
-    int semi, area;
+    float semi, area;
     printf("Ingrese un lado del triangulo: ");
     scanf("%d",&a);
     printf("Ingrese un lado del triangulo: ");
@@ -11,10 +11,10 @@ int main() {
     printf("Ingrese un lado del triangulo: ");
     scanf("%d",&c);
     if (a + b > c && a + c > b && b + c > a) {
-        semi = (a + b + c) / 2;
+        semi = (a + b + c) / 2.0;
         area = sqrt(semi * (semi - a) * (semi - b) * (semi - c));
-        printf("El semiperimetro es: %d\n",semi);
-        printf("El area es: %d\n",area);
+        printf("El semiperimetro es: %f\n",semi);
+        printf("El area es: %f\n",area);
     }
     else {
         printf("Triangulo no valido\n");
